move runbigscape.py to bigscape folder
