+++ src/nplinker/arranger.py
@@ -14,7 +14,7 @@ from nplinker.globals import STRAIN_MAPPINGS_FILENAME
 from nplinker.metabolomics.gnps import GNPSDownloader
 from nplinker.metabolomics.gnps import GNPSExtractor
 from nplinker.pairedomics import podp_download_and_extract_antismash_data
-from nplinker.pairedomics.runbigscape import run_bigscape
+from nplinker.genomics.bigscape.runbigscape import run_bigscape
 from nplinker.pairedomics.strain_mappings_generator import podp_generate_strain_mappings
 from nplinker.schemas import STRAIN_MAPPINGS_SCHEMA
 from nplinker.schemas import USER_STRAINS_SCHEMA
