+++ src/nplinker/genomics/bigscape/__init__.py
@@ -1,7 +1,8 @@
 import logging
 from .bigscape_loader import BigscapeGCFLoader
+from .runbigscape import run_bigscape
 
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
-__all__ = ["BigscapeGCFLoader"]
+__all__ = ["BigscapeGCFLoader", "run_bigscape"]
